imnetio: unit test sending request with headers
